Include bag information as part of entity state
